MES-Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==3.0.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2021.1
requests==2.32.3
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
